Move Call Log backup into it's own package.
Moves the call log backup from the contacts provider into it's own
package.  This should make it easier to support backup through the same
package name for OEMs which revise (and rename) the contacts provider.

Also add a CallLogChangeReceiver to receive notices from the
CallLogProvider when the call log changes.  When this happens, the
receiver notifies the backup manager that it should run a backup at it's
earliest convenience.

Change-Id: I5774bba333a83adf53e1733f80a3df9e8140adb8