Mark missed calls as read in calllog restore(1/2)
When missed calls are restored by CalllogBackupAgent using the addCall
API of CallLog provider, we were setting the IS_READ flag to 0 which
resulted in user getting a missed call notifications for all missed
calls after restore.
Adding a new addCall API signature which can specify the value to be put
in the
IS_READ field when a call is added to the calllog provider. This will be
used by the CallLogBackupAgent to mark all restored missed calls as
read.

BUG: 22908952
Change-Id: Ie8e6619b75801c8f0ba4f554a5c3e67a7ad706dd