Catch all types of exceptions in CallLogBackupAgent#runBackup.
By the current implementation, CallLogBackupAgent#runBackup could be invoked with a null value passed in for BackupDataOutput. This would result in a fatal NPE that is not currently being caught or handled anywhere. There is no non-null check at a higher level (such as BackupAgent#onBackup) to prevent this issue either. This CL ensures that all types of exceptions are caught and handled correctly.

Fixes: 279649049
Test: atest CallLogBackupAgentTest#testRunBackup_OneNewCall_NullBackupDataOutput
Change-Id: Ibddbb86b97ff5e00ee2e31527df7a1a0e5bdcbda